// open-api-modules/customer-api-docs/model/apiResponseCustomerInfoResponse.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomerInfoResponse } from './customerInfoResponse';


export interface ApiResponseCustomerInfoResponse { 
    result?: CustomerInfoResponse;
    errorCode?: string;
    message?: object;
    responseCode?: number;
}